Replace `as_*` TokenIterExt methods with `try_*`
The `as_*` methods always consume the item from the token stream, even if the item doesn't match the kind of token requested. This makes it very difficult for parsers to try multiple potential tokens.

This replaces these methods with "non-consuming" implementations that will leave the underlying iterator untouched when they return an error. This is a breaking change anyway, so the methods have been renamed to better reflect what they do.

- Adds tests
- Closes #7